separate logging target test function to different file
Problem:
file testwrap.go include different target test function, too long

Solution:
separate logging test function to their own file, this commit just separate code, not change logic, not do this in previous commit because it will hard to review code
