+++ pkg/controllers/user/logging/utils/templatewrap.go
@@ -296,15 +296,3 @@ func parseEndpoint(endpoint string) (host string, scheme string, err error) {
 
 	return u.Host, u.Scheme, nil
 }
-
-func getDateFormat(dateformat string) string {
-	ToRealMap := map[string]string{
-		"YYYY-MM-DD": "%Y-%m-%d",
-		"YYYY-MM":    "%Y-%m",
-		"YYYY":       "%Y",
-	}
-	if _, ok := ToRealMap[dateformat]; ok {
-		return ToRealMap[dateformat]
-	}
-	return "%Y-%m-%d"
-}
